+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Fixes
 
 * Fix early return in FindMessageByName (#2644)
+* Preserve HTTP status for retriable SR errors (#2647)
 
 
 # 2.15.0